+++ src/Universum/Lifted.hs
@@ -7,11 +7,9 @@ module Universum.Lifted
        , module Universum.Lifted.Env
        , module Universum.Lifted.File
        , module Universum.Lifted.IORef
-       , module Universum.Lifted.ST
        ) where
 
 import Universum.Lifted.Concurrent
 import Universum.Lifted.Env
 import Universum.Lifted.File
 import Universum.Lifted.IORef
-import Universum.Lifted.ST