+++ src/Universum/Function.hs
@@ -2,13 +2,6 @@
 
 module Universum.Function
        ( module Data.Function
-       , identity
        ) where
 
 import Data.Function (const, fix, flip, id, on, ($), (.))
-
--- | Renamed version of 'Prelude.id'.
-identity :: a -> a
-identity = id
-{-# INLINE identity #-}
-{-# DEPRECATED identity "Use `id` instead" #-}