+++ CHANGES.md
@@ -1,3 +1,32 @@
+1.4.0
+=====
+* [#167](https://github.com/serokell/universum/issues/164):
+  `identity` has been removed.
+
+  _Migration guide:_ use `Universum.id` instead.
+* [#177](https://github.com/serokell/universum/issues/177):
+  The `mask_` reexport from `safe-exceptions` has been removed.
+
+  _Migration_guide:_ use `Control.Exception.Safe.mask_` from `safe-exceptions`
+  instead.
+* [#178](https://github.com/serokell/universum/issues/178):
+  `getArgs` has been removed.
+
+  _Migration guide:_ use `liftIO` directly with `System.Environment.getArgs`
+  from base.
+* [#179](https://github.com/serokell/universum/issues/179):
+  `getContents` and `interact` have been removed.
+
+  _Migration guide:_ use `liftIO` directly with `Data.Text.Lazy.IO.getContents`
+  and `Data.Text.Lazy.IO.interact`, both from the `text` package.
+* [#180](https://github.com/serokell/universum/issues/180):
+  The `Lifted.ST` module has been removed.
+  
+  _Migration guide:_ use `liftIO` directly with functions from
+  `Control.Monad.ST` instead.
+* [#181](https://github.com/serokell/universum/issues/181):
+  `list` has removed.
+
 1.3.0
 =====